requirements.txt
streamlit
openai
SQLAlchemy
pandas
icalendar
PyPDF2
python-dotenv
